extracted compression modules
